+++ CHANGELOG.md
@@ -158,14 +158,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.1] - 2019-01-20
 ### Added
  - Custom Proton installations in `STEAM_DIR/compatibilitytools.d` are now detected. See [Sirmentio/protontricks#31](https://github.com/Sirmentio/protontricks/issues/31).
- - protontricks is now a Python package and can be installed using `pip`.
+ - Protontricks is now a Python package and can be installed using `pip`.
  
 ### Changed
  - Argument parsing has been refactored to use argparse.
    - `protontricks gui` is now `protontricks --gui`.
    - New `protontricks --version` command to print the version number.
  - Game names are now displayed in alphabetical order and filtered to exclude non-Proton games.
- - protontricks no longer prints INFO messages by default. To restore previous behavior, use the `-v` flag.
+ - Protontricks no longer prints INFO messages by default. To restore previous behavior, use the `-v` flag.
 
 ### Fixed
  - More robust VDF parsing.
@@ -174,7 +174,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## 1.0 - 2019-01-16
 ### Added
- - The last release of protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
+ - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
 [Unreleased]: https://github.com/Matoking/protontricks/compare/1.6.2...HEAD
 [1.6.2]: https://github.com/Matoking/protontricks/compare/1.6.1...1.6.2