+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
-## [Unreleased]
+## [1.7.0] - 2022-01-08
 ### Changed
  - Enable usage of Flatpak Protontricks with non-Flatpak Steam. Flatpak Steam is prioritized if both are found.
 
@@ -177,7 +177,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
  - The last release of Protontricks maintained by [@Sirmentio](https://github.com/Sirmentio).
 
-[Unreleased]: https://github.com/Matoking/protontricks/compare/1.6.2...HEAD
+[Unreleased]: https://github.com/Matoking/protontricks/compare/1.7.0...HEAD
+[1.7.0]: https://github.com/Matoking/protontricks/compare/1.6.2...1.7.0
 [1.6.2]: https://github.com/Matoking/protontricks/compare/1.6.1...1.6.2
 [1.6.1]: https://github.com/Matoking/protontricks/compare/1.6.0...1.6.1
 [1.6.0]: https://github.com/Matoking/protontricks/compare/1.5.2...1.6.0